+++ src/components/ViewSource/ViewSource.js
@@ -18,7 +18,10 @@ import {
 
 import { useGoBack } from '../../common/hooks';
 
-import { isUserInConsortiumMode } from '../../utils';
+import {
+  isInstanceShadowCopy,
+  isUserInConsortiumMode,
+} from '../../utils';
 import MARC_TYPES from './marcTypes';
 
 import styles from './ViewSource.css';
@@ -88,7 +91,7 @@ const ViewSource = ({
     <FormattedMessage
       id={`ui-inventory.marcSourceRecord.${marcType}`}
       values={{
-        shared: isUserInConsortiumMode(stripes) ? instance.shared : null,
+        shared: isUserInConsortiumMode(stripes) ? (instance.shared || isInstanceShadowCopy(instance?.source)) : null,
       }}
     />
   );